refactor: AI Code Review 리팩토링
