+++ src/main/java/com/aibe/team2/domain/interview/repository/InterviewSessionRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.aibe.team2.domain.interview.repository;
 
-import com.aibe.team2.domain.mypage.dto.InterviewSessionListResponse;
+import com.aibe.team2.domain.mypage.dto.response.InterviewSessionListResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
